Devo to Main

+++ app/user/__init__.py
@@ -1,4 +1,4 @@
-from flask import render_template, Blueprint, redirect, url_for
+from flask import render_template, Blueprint, redirect, url_for, abort
 from flask_login import login_required, current_user
 
 from app.auth.decorators import admin_required
@@ -9,12 +9,12 @@ from app.user.forms import EditForm, DeleteForm
 users = Blueprint("users", __name__, template_folder="templates")
 
 
-@users.route("/users/edit/", methods=["GET", "POST"])
+@users.route("/users/edit", methods=["GET", "POST"])
 @login_required
 def edit():
     edit_form = EditForm()
 
-    user = User.query.filter(User.username == current_user.username).first()
+    user = User.query.filter_by(username=current_user.username).first()
 
     if user:
         if edit_form.validate_on_submit():
@@ -37,14 +37,17 @@ def edit():
 
 
 @users.route("/users/delete", methods=["POST"])
+@login_required
 @admin_required
 def delete():
     delete_form = DeleteForm()
 
     if delete_form.validate_on_submit():
         user = User.query.filter_by(username=delete_form.username.data).first()
-        if user:
+        if user and not user.is_admin:
             db.session.delete(user)
             db.session.commit()
+        else:
+            abort(400, "An Admin cannot be deleted!")
 
     return redirect(url_for("admin.index"))
